+++ src/components/Slider/locale/en_US.js
@@ -7,6 +7,6 @@ export default {
     comma: ', ',
     input: 'input',
     isValid: ' is valid',
-    isUnvalid: ' is unvalid',
+    isInvalid: ' is invalid',
     tip: 'please press enter or click other place to accept valid value'
 };